+++ groups/bae/baem/baem_publicationtype.h
@@ -182,7 +182,9 @@ int baem_PublicationType::maxSupportedBdexVersion()
 inline
 int baem_PublicationType::fromString(Value *result, const bsl::string& string)
 {
-    return fromString(result, string.c_str(), string.length());
+    return fromString(result,
+                      string.c_str(),
+                      static_cast<int>(string.length()));
 }
 
 inline
